Moved the content of keycode.h to input.h
